Added compress & backup Saves directory functionality
Added the ability to compress and backup the Saves directory to the command line, settings file, and automatically during the update all process. Also added a setting to allow an absolute path for the zip file to be created on rather than the local directory where pupdate is being run from.

Corrected a spelling mistake with _downloadFirmware in the Updater.cs file.